refactor: make clear cache an independent component
