+++ packages/neuron-ui/src/components/ClearCache/style.module.scss
@@ -1,0 +1,26 @@
+.clearCache {
+  &.detail {
+    grid-area: clear-cache-detail;
+    .date {
+      display: flex;
+      align-items: center;
+      font-size: 0.875rem;
+      height: 1.125rem;
+      margin-bottom: 5px;
+    }
+    .desc {
+      display: flex;
+      font-size: 0.6875rem;
+      color: #666;
+    }
+    svg {
+      height: 0.6875rem;
+      width: 0.6875rem;
+      filter: grayscale(1) opacity(0.6);
+      margin: 2px 5px 0 0;
+    }
+  }
+  &.action {
+    grid-area: clear-cache-action;
+  }
+}
